Safari-611.1.10/LayoutTests/http/wpt/webauthn/public-key-credential-create-success-u2f.https-expected.txt

CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.create' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.create' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.create' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.create' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.create' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.create' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.create' within user activated events.

PASS PublicKeyCredential's [[create]] with minimum options in a mock u2f authenticator.
PASS PublicKeyCredential's [[create]] with excludeCredentials in a mock u2f authenticator.
PASS PublicKeyCredential's [[create]] with excludeCredentials in a mock u2f authenticator. 2
PASS PublicKeyCredential's [[create]] with test of user presence in a mock u2f authenticator.
PASS PublicKeyCredential's [[create]] with none attestation in a mock u2f authenticator.
PASS PublicKeyCredential's [[create]] with indirect attestation in a mock u2f authenticator.
PASS PublicKeyCredential's [[create]] with direct attestation in a mock u2f authenticator.
